requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-data-seeder==0.2.0
django-filter==21.1
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.3.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.9
